+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.7 (unreleased)
 
+- Added `Upsample`
 - Fixed error with buffers on GPU
 
 ## 0.3.6 (2020-09-17)