+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-## 0.3.7 (unreleased)
+## 0.3.7 (2020-09-22)
 
+- Improved performance
 - Added `Upsample`
 - Added support for passing tensor class to `type` method
-- Improved performance
 - Fixed error with buffers on GPU
 - Fixed error with `new_full`
 - Fixed issue with `numo` method and non-contiguous tensors